% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figpatch.R
\name{img2plot}
\alias{img2plot}
\title{img2plot
Convert image to ggplot object with given aspect ratio.}
\usage{
img2plot(path, AR = "default", colour = NULL, fill = NULL)
}
\arguments{
\item{path}{Path to image file.}

\item{AR}{Aspect ratio. 'default' inherits image's aspect ratio.}

\item{colour}{Colour of rectangle around image. (Uses plot.background in theme)}

\item{fill}{Fill of rectange around image. (uses plot.background in theme)}
}
\value{

}
\description{
img2plot
Convert image to ggplot object with given aspect ratio.
}
