% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figpatch.R
\name{plotlab}
\alias{plotlab}
\title{Label img2plot plot.}
\usage{
plotlab(
  plot,
  lab,
  pos = "topleft",
  colour = NULL,
  alpha = NULL,
  hjust = 0,
  vjust = 1,
  fontsize = 20,
  fontface = NULL,
  fontfamily = NULL
)
}
\arguments{
\item{plot}{Plot from img2plot function.}

\item{lab}{Label to add to plot.}

\item{pos}{Position of label (Default 'topleft').}

\item{colour}{Colour of label text.}

\item{alpha}{Alpha of label text.}

\item{hjust}{hjust of plot label.}

\item{vjust}{vjust of plot label.}

\item{fontsize}{Fontsize of laebl (in points).}

\item{fontface}{The font face (bolt, italic, ...)}

\item{fontfamily}{Fontfamily of plot label.}
}
\value{

}
\description{
Label img2plot plot.
}
