% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figpatch.R
\name{fig}
\alias{fig}
\title{fig
Read image and convert to ggplot object with given aspect ratio.}
\usage{
fig(path, AR = "default", colour = NULL, fill = NULL)
}
\arguments{
\item{path}{Path to image file.}

\item{AR}{Aspect ratio. 'default' inherits image's aspect ratio.}

\item{colour}{Colour of rectangle around image. (Uses plot.background in theme)}

\item{fill}{Fill of rectangle around image. (uses plot.background in theme)}
}
\value{

}
\description{
fig
Read image and convert to ggplot object with given aspect ratio.
}
