% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.R
\name{fig}
\alias{fig}
\title{Create a fig.
Read image and convert to ggplot object, for use with other ggplot objects
when assembling with the \code{{patchwork}} package. Can also specify a border.}
\usage{
fig(
  path,
  aspect.ratio = "default",
  b_col = NULL,
  b_size = 1,
  b_pos = "offset",
  b_margins = 0.01,
  b_unit = "npc"
)
}
\arguments{
\item{path}{Path to image file.}

\item{aspect.ratio}{Manually override the image's aspect ratio or set "free" to allow fig
to be resized by patchwork.}

\item{b_col}{Colour of the border line.}

\item{b_size}{Size of the border line.}

\item{b_pos}{Whether the border is 'offset' (expands out from figure) or
'inset' and expands inwards, partially covering up the figure.}

\item{b_margins}{Margins around the fig. Dimensions for top, right, bottom
and left margins. Numeric vector of 1, 2 will be recycled for all 4
dimensions.}

\item{b_unit}{Unit used for margins. Defaults to 'npc'. (see: ggplot2::unit())}
}
\value{
ggplot2 plot with external figure.
}
\description{
Create a fig.
Read image and convert to ggplot object, for use with other ggplot objects
when assembling with the \code{{patchwork}} package. Can also specify a border.
}
