% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figlab.R
\name{figlab}
\alias{figlab}
\title{Add tags and a caption to a fig.}
\usage{
figlab(
  plot,
  tag,
  pos = "topleft",
  x_nudge = 0,
  y_nudge = 0,
  colour = NULL,
  alpha = NULL,
  hjust = NULL,
  vjust = NULL,
  fontsize = 12,
  fontface = NULL,
  fontfamily = NULL
)
}
\arguments{
\item{plot}{Plot from img2plot function.}

\item{tag}{Label to add to plot.}

\item{pos}{Position of label (Default 'topleft').}

\item{x_nudge}{Minor adjustments to the x position in relative plot
coordindates (0 being furthest left, 1 being furthest right).}

\item{y_nudge}{Minor adjustments to the y position in relative plot
coordinates (0 being the bottom, 1 being the top).}

\item{colour}{Colour of label text.}

\item{alpha}{Alpha of label text.}

\item{hjust}{hjust of plot label.}

\item{vjust}{vjust of plot label.}

\item{fontsize}{Fontsize of laebl (in points).}

\item{fontface}{The font face (bolt, italic, ...)}

\item{fontfamily}{Fontfamily of plot label.}
}
\value{
ggplot2 fig with label added.
}
\description{
Add tags and a caption to a fig.
}
