% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figpatch.R
\name{figwrap}
\alias{figwrap}
\title{Quickly arrange and label multiple figs.}
\usage{
figwrap(
  figs,
  labelling = NULL,
  prefix = NULL,
  suffix = NULL,
  pos = "topleft",
  nrow = NULL,
  ncol = NULL,
  colour = NULL,
  alpha = NULL,
  hjust = NULL,
  vjust = NULL,
  fontsize = NULL,
  fontfamily = NULL,
  fontface = NULL
)
}
\arguments{
\item{figs}{List of figs from \code{fig()}.}

\item{labelling}{}

\item{prefix}{Prefix for each label.}

\item{suffix}{Suffix for each label.}

\item{pos}{Position for label, to be passed to \code{figlab()}.}

\item{nrow}{Number of rows in final patchwork.}

\item{ncol}{Number of cols in final patchwork.}

\item{colour}{Colour for each label.}

\item{alpha}{Alpha for each label.}

\item{hjust}{hjust for each label.}

\item{vjust}{vjust for each label.}

\item{fontsize}{Fontsize for each label.}

\item{fontfamily}{Fontfamily for each label.}

\item{fontface}{Fontface for each label.}
}
\value{

}
\description{
Quickly arrange and label multiple figs.
}
