% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figwrap.R
\name{figwrap}
\alias{figwrap}
\title{Quickly arrange and label multiple figs.}
\usage{
figwrap(
  figs,
  tag = NULL,
  prefix = NULL,
  suffix = NULL,
  pos = "topleft",
  x_nudge = 0,
  y_nudge = 0,
  nrow = NULL,
  ncol = NULL,
  colour = NULL,
  alpha = NULL,
  hjust = NULL,
  vjust = NULL,
  fontsize = NULL,
  fontfamily = NULL,
  fontface = NULL,
  b_col = NULL,
  b_size = 1,
  b_pos = "offset",
  b_margins = NULL,
  b_unit = NULL
)
}
\arguments{
\item{figs}{List of figs from \code{fig()}.}

\item{tag}{Tags to be applied to figs. Begins with first uppercase or
lowercase letter supplied, or number, and continues the sequence. "A"
labels them 'A', 'B', etc. "c" labels them 'c', 'd', 'e' etc.}

\item{prefix}{Prefix for each tag}

\item{suffix}{Suffix for each tag}

\item{pos}{Position for label, to be passed to \code{figlab()}.}

\item{x_nudge}{Minor adjustments to the x position in relative plot
coordindates (0 being furthest left, 1 being furthest right).}

\item{y_nudge}{Minor adjustments to the y position in relative plot
coordinates (0 being the bottom, 1 being the top).}

\item{nrow}{Number of rows in final patchwork.}

\item{ncol}{Number of cols in final patchwork.}

\item{colour}{Colour for each tag}

\item{alpha}{Alpha for each tag}

\item{hjust}{hjust for each tag}

\item{vjust}{vjust for each tag}

\item{fontsize}{Fontsize for each tag}

\item{fontfamily}{Fontfamily for each tag}

\item{fontface}{Fontface for each tag}

\item{b_col}{Colour of individual fig borders.}

\item{b_size}{Size of individual fig borders (in mm).}

\item{b_pos}{Either "offset" and expanding outwards from borders of figure,
or "inset" and expanding inwards and partially covering the figure.}

\item{b_margins}{Adjust margins for the plots, in given units (default
"npc"). c(top, right, bottom, left) margins to be supplied, if vector of
length 1 or 2 is given, values will be recycled to to make a vector of
length 4.}

\item{b_unit}{Unit for the margin adjustment, defaults to "npc".}
}
\value{
\code{patchwork} patch of supplied figs.
}
\description{
Quickly arrange and label multiple figs.
}
\examples{

library(figpatch)
library(ggplot2)

# Attach the fig image file
image <- system.file("extdata", "fig.png", package = "figpatch", mustWork = TRUE)

# Read in the image as a 'fig'
img <- fig(image)

# multiple figs
figs <- lapply(1:9, function(x) img)

# wrap the figs
figwrap(figs)

# Wrap the figs and auto-tag
figwrap(figs, tag = "A", suffix = ")")

# Wrap figs, auto-tag and adds border.
figwrap(figs, tag = 1, prefix = "(", suffix = ")", b_col = "black")
}
